+++ apps/server/src/middleware/auth.middleware.ts
@@ -3,7 +3,7 @@ import jwt from 'jsonwebtoken'
 
 import { env } from '../env'
 
-import type { TokenPayload } from '../modules/auth/auth.utils'
+import type { TokenPayload } from '../modules/auth/auth.types'
 import type { Request, Response, NextFunction } from 'express'
 
 export const authMiddleware = async (req: Request, res: Response, next: NextFunction) => {
