--- packages/core/package.json
@@ -1,14 +1,0 @@
-{
-    "name": "@december/core",
-    "module": "index.ts",
-    "main": "./index.ts",
-    "types": "./index.ts",
-    "type": "module",
-    "private": true,
-    "devDependencies": {
-        "@types/bun": "latest"
-    },
-    "peerDependencies": {
-        "typescript": "^5"
-    }
-}
